Change servername for *.incubator.a.o
It appears that SNI on the AWS host isn't working as expected and
is using the *.apache.org cert for *.incubator.a.o requests.
Changing the priority for incubator vhost only reverses the problem.

Forcing the incubator vhost ServerName to be incubator.apache.org
allows *.incubator.apache.org requests to work properly as well as
*.apache.org requests. The only downside is that incubator.apache.org
uses the *.incubator.apache.org cert, and therefore on browsers it will
not be able to supply site identity information.