convert %(...) => %{...} in vhost custom_fragments
Background: hiera and Apache unfortunately have the same escape sequence for
interpolating variables in strings.  While there are various workarounds for
escaping the escape sequence (%%{}{var} and %{literal('%'}}{var} are both
popular), neither are working for me when trying to do <If> statements.

After more than an hour of trying various alternatives, it seems best to
avoid the syntax clash altogether.  As this change will need to be deployed
and the puppet master restarted before it can be used, I'm pushing it
separately.