allow LDAP attribute to be specified in authorization templates
look for ";attr=" within {ldap:} sections, and extract attribute name
if this string is present; otherwise default to memberUid,member for
backwards compatibility.

A single example of this new syntax is included in this commit.

The plans are to shortly make use of this to move PMC authorizations to
an owner attribute.