mysql rsync
add a flaggable ability to push mysql db dumps offsite to bai via rsync.
add stunnel_asf as a dependency to mysql_asf::backup. This shouldn't
affect machines that don't need it, but is necessary to make offsite
backups use only a single flag. nb: this will also require further
commits/changes to enable the backup on bai and eyaml to add the rsync
passwords.